pyproject.toml
[project]
name = "anai"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ronak Mehta", email = "ronakrm@gmail.com" }
]
dependencies = [
    "anthropic>=0.34.2",
    "pandas>=2.2.3",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    # Add this line
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["anai"]

[tool.pytest.ini_options]
markers = [
    "live_api: marks tests that use the live API (deselect with '-m \"not live_api\"')"
]
